@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,14 +27,16 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
-  curved_navigation_bar: ^0.3.4
-  google_maps_flutter: ^0.5.33
-  geolocator: ^6.0.0+4
+  curved_navigation_bar: ^0.4.0-nullsafety.0
+  google_maps_flutter: ^2.0.3
+  geolocator: ^7.0.3
   firebase_core: ^0.4.0+9
   firebase_analytics: ^5.0.2
   firebase_auth: ^0.15.3
-  fluttertoast: ^7.1.1
-  flutter_facebook_login: ^3.0.0
+  fluttertoast: ^8.0.6
+  flutter_facebook_auth: ^3.3.3
+#  flutter_facebook_login: ^3.0.0
+#  curved_navigation_bar: ^0.3.7
 
 dev_dependencies:
   flutter_test:
